requirements.txt
docker-py
argparse
urllib3
chardet
requests